+++ utilities/transform_data_stations_v0.3.0.py
@@ -98,7 +98,7 @@ for year in range(years):
                                             "(threshold below [default or personal threshold])",
                                             "(threshold above [default or personal threshold])",
                                             "(analysis interval [5days,10days,15days or trimester])",
-                                            "(forecasting values: 9 values)", "(forecasting date [trimester or month/day])"])
+                                            "(forecast values: 9 values)", "(forecast date [trimester or month/day])"])
 
         for month in range(12):